app/requirements.txt
Django==2.2
psycopg2==2.8
djangorestframework==3.12.4